app/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
streamlit
requests